# api/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API voyage. Documentation",
        "description": "Cette API permet aux tours-opérateurs de gérer les avis de leurs clients sur les voyages qu'ils proposent.\r\n\r\n    ",
        "contact": {
            "name": "the developer",
            "email": "sengerromain75@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1/php/agence_voyage/api/",
            "description": "Serveur local de développement"
        }
    ],
    "paths": {
...
        "/avis/create": {
            "post": {
                "tags": [
                    "Avis"
                ],
                "summary": "Ajouter un avis",
                "description": "Ajouter un avis",
                "operationId": "45721ba52b485f856844c9dcdfe7c45f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "avis",
                                    "voyageID",
                                    "clientID"
                                ],
                                "properties": {
                                    "avis": {
                                        "description": "Avis du client",
                                        "type": "string",
                                        "example": "Avis client"
                                    },
                                    "voyageID": {
                                        "description": "ID du voyage associé à l'avis",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "clientID": {
                                        "description": "ID du client associé à l'avis",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Inserer les datas"
                    },
                    "400": {
                        "description": "Tous les champs sont obligatoires"
                    },
                    "405": {
                        "description": "Méthode non autorisée"
                    }
                }
            }
        },
        "/avis/read/{avisID}": {
            "get": {
                "tags": [
                    "Avis"
                ],
                "summary": "Lire un avis",
                "description": "Lire un avis",
                "operationId": "fd70686119bc63ccc1df50d658920b1d",
                "parameters": [
                    {
                        "name": "avisID",
                        "in": "path",
                        "description": "avisID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Avis trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Aviss"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Avis non-trouvé"
                    },
                    "405": {
                        "description": "Méthode non-autorisé"
                    }
                }
            }
        },
        "/avis/readAll": {
            "get": {
                "tags": [
                    "Avis"
                ],
                "summary": "Afficher tous les avis",
                "description": "Lire tous les avis",
                "operationId": "36b030ee45ab361a735fdbc1f30f453e",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Aviss"
                                    }
                                }
                            }
                        }
                    },
                    "405": {
                        "description": "Mauvaise Méthode"
                    }
                }
            }
        },
        "/avis/update": {
            "put": {
                "tags": [
                    "Avis"
                ],
                "summary": "Modifier un client",
                "description": "Modifier un avis",
                "operationId": "141a8dde69ed5fe7982f98b6c6b4798f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "avis",
                                    "voyageID",
                                    "clientID",
                                    "avisID"
                                ],
                                "properties": {
                                    "avis": {
                                        "description": "Avis du client",
                                        "type": "string",
                                        "example": "Avis client"
                                    },
                                    "voyageID": {
                                        "description": "ID du voyage associé à l'avis",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "clientID": {
                                        "description": "ID du client associé à l'avis",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "avisID": {
                                        "description": "ID de l'avis",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Modifier les datas"
                    },
                    "400": {
                        "description": "Tous les champs sont obligatoires"
                    },
                    "405": {
                        "description": "Méthode non autorisée"
                    }
                }
            }
        },
        "/avis/delete": {
            "delete": {
                "tags": [
                    "Avis"
                ],
                "summary": "Supprimer un Avis",
                "description": "Supprimer un avis",
                "operationId": "83ee9df6754ea6237c69088f782e0857",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "avisID"
                                ],
                                "properties": {
                                    "avisID": {
                                        "description": "ID de l'avis à supprimer",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Supprimer les datas"
                    },
                    "400": {
                        "description": "Tous les champs sont obligatoires"
                    },
                    "405": {
                        "description": "Méthode non autorisée"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Aviss": {
                "description": "Affichage d'un avis client",
                "properties": {
                    "avisID": {
                        "description": "Clé unique de l'avis",
                        "type": "integer"
                    },
                    "avis": {
                        "description": "avis du client sur le voyage",
                        "type": "string"
                    },
                    "voyageID": {
                        "description": "Identifiant du voyage",
                        "type": "integer"
                    },
                    "clientID": {
                        "description": "Identifiant du client",
                        "type": "integer"
                    },
                    "voyage": {
                        "description": "Détails du voyage",
                        "properties": {
                            "titre": {
                                "description": "Titre du voyage",
                                "type": "string"
                            },
                            "description": {
                                "description": "Description du voyage",
                                "type": "string"
                            }
                        },
                        "type": "object"
                    },
                    "client": {
                        "description": "Détails du client",
                        "properties": {
                            "prenom": {
                                "description": "Prénom du client",
                                "type": "string"
                            },
                            "nom": {
                                "description": "Nom du client",
                                "type": "string"
                            },
                            "email": {
                                "description": "Email du client",
                                "type": "string"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}